requirements_compile_win311.txt
# Minimal-ish MESA dev+compile environment for Windows + Python 3.11
# Focus: run scripts + build executables via PyInstaller (devtools/build_all.py / devtools/compile_win_11.bat)
# If you need every optional tool/package, use requirements_all_win311.txt instead.

# Build toolchain
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1

# GUI
ttkbootstrap==1.10.1
pywebview

# Core numeric/data stack
numpy==1.26.2
pandas==2.1.4
pyarrow==17.0.0

# Geospatial stack (used by most tools)
shapely==2.0.2
pyproj==3.6.1
fiona==1.10.1
geopandas==1.1.2
pyogrio==0.7.2
rasterio==1.3.9
Rtree==1.2.0

# Common utilities used across scripts
requests==2.32.4
Pillow==12.1.1
openpyxl==3.1.2
XlsxWriter==3.1.9
reportlab==4.2.0
tqdm==4.67.1
python-docx==1.2.0
lxml==6.0.2

# setuptools/pkg_resources runtime deps (needed in frozen builds)
jaraco.text==4.0.0
jaraco.functools==4.4.0
jaraco.context==6.1.0
more-itertools==10.8.0
platformdirs==4.5.1
autocommand==2.2.2
backports.tarfile==1.2.0

# Some scripts reference H3
h3==4.2.2
